Added some files and reorganized a bit
Folders where made to know where to find stuff and added some literature
that will be useful when preparing for exam. + i kind of started a
report and saved in two type of documents so i could send it to Henrik
and he could give us some pointers for it.